lambda_analyst/magicmirror/requirements.txt
# CodeRipple Requirements
# Core Strands framework and tools
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Standard library extensions (these are built-in but listing for clarity)
# pathlib - built-in
# subprocess - built-in
# logging - built-in
# os - built-in
# sys - built-in

# For development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# For AWS Lambda deployment (optional - only needed for Lambda)
boto3>=1.26.0
botocore>=1.29.0

# For enhanced git operations (optional - git command line tools should be available)
# GitPython>=3.1.0  # Uncomment if you want Python git interface instead of shell commands